(maint) Do not symbolize the resultset hash
Previous to this commit, we were symbolizing all keys in the hash being
returned to puppet. However, due to how puppet transforms the hash into
objects for immutability, the keys were not being picked up.
This commit converts everything to string based to allow consumers to
actually use the return resultset