Make run_puppet() environment parameter optional
Previous to this commit, the run_puppet() function's environnment
parameter was required. However, puppet agents already have a default
environmnent assigned to them either through the puppet.conf file or an
environment node group. This option should only be required when the
user wants to override the agent's assigned environment (and the agent
is permitted to do so)