Remove unnecessary git branch deletions
Previous to this commit, the rolling deployment policy would try to
ensure that the temporary git branch created in the deployment is
removed. However, that's not necessary due to the create_git_branch
defaulting to deleting itself after the deployment is complete, even if
the deployment fails.

This commit also reimplements the code deployment validation method