Support prefixed environments
Previous to this commit, the update_git_branch function targeted the
branch of the same name as the target environment. However, that won't
work with prefixed environments because the environment name will
include the prefix but the branch name will not. Therefor, this commit
uses the BRANCH environment variable instead.