CDPE-2650 create module branch if missing on eventual consistency
Currently, we try to update the git ref of the entity in question on an eventual consistency deployment, no questions asked. However, in the case of a module, the branch may not exist as it comes from a node group name not an actual branch. To get back to parity with PE Module Deployments, this change adds a check to see if the update ref fails and the repo type is a module. If so, try to create a branch from the src sha instead.