Don't fail if module branch doesn't exist
Previous to this commit, when an eventual consistency deployment
policy was run in a module pipeline and the target branch didn't exist
on the module's vcs repository, the deployment would successfully create
the target branch on the module's repository, but fail the deployment. A
follow up deployment would succeed.

This commit ensures the branch exists before performing an update_ref by
calling cd4pe_deployments::create_git_branch() first. This is a safe
operation because create_git_branch() is idempotent. i.e. it will
succeed if the branch already exists.