CDPE-2695: Update CD4PE bolt job to be a script
With this change, we modify the Bolt task we wrote to execute CD4PE jobs
to be a "script" in order for it to be executed via the Puppet
Orchestrator + Puppet Communications Protocol (tasks can not be
executed via PCP unless the task is present on the PE environment
itself. By using 'run_script' instead of 'run_task', we circumvent
this issue).

To make this a script, we had to consolidate a bunch of classes into a
single file. In order to maintain testability, those classes are
all defined in the new single file.

Unit tests have been combined and modified as well to validate all the
new logic we have added here.