CDPE-3014 gets cd4pe_token from system env
Currently, when viewing a run of the cd4pe_job plan in the PE console, you can see the '$cd4pe_token' value plain as day. While this token only lasts the life of the task run, there is a small chance it could be used for something other than intended so we are changing the way we get it. Instead of it being passed to the plan as a param, we're going to set it in CD4PE and then get it from the environment in our plan. This change gets it using system::env().