(CDPE-2789) Add a function to list branches for deployment repo
- Previously, users could create and delete branches via the deployment
related Puppet functions, but there was no way to list available branches
 for a given repository associated with a deployment.
- Adds a new function to allow users to list git branches for a repo
associated with a deployment. It's possible to list branches for the
control repo or module repo. The reason this is specified at all is
because a module deployment is associated with a repository, but can
also be associated with a control repo which is why the function takes
the repo type as an argument.
- In addition to the tests added here, The underlying CD4PE API has
 been manually tested with every  VCS provider.