(CDPE-3193) Perform a code deploy after rolling deployment
Previous to this commit, after a rolling deployment is completed, we
would perform a merge and delete the temp environment, but not deploy
the final merged code. This would cause the agents on their next run to
revert to the previous code, causing production outages.

This commit fixes the issue by performed the code deploy right after the
code is merged, but before any other operations to reduce risk of
failures leaving code un-deployed.