Prevent repeat deployment approval pending states
Previous to this commit, the cd4pe_deployments::wait_for_approval() function
would immediately attend to put the deployment approval into a pending
state. However, it's possible for the approval to already be approved
earlier in the deployment plan, before the wait_for_approval() function
is called.

This commit first checks the status of the deployment approval and
returns with the status if it's already approved or denied