(CDPE-3405) Lowercase rolling temp branch name
Previuos to this commit, our temp branch for rolling plan was created
with all caps `ROLLING_DEPLOYMENT`. However, puppet has a suggested
regex that all branch names should match (lowercase numbers and 0-9).
This commit changes the prefix we add to be lowercase. Since it is just
a suggestion, and depending on OS git branches are case sensitive, it is
best to not do a global downcase to not potentially break customer
workflows that depend on case sensitivity.