(CDPE-3788) Use env var for node group env instead of api
Previous to this commit, the eventual consistency plan would query a
CD4PE api endpoint that did recursive nodegroup rules lookups that
returned all nodes in the target environment. This was causing issues
for customers with large, complex rule sets. Eventual Consistency
doesn't care about the node list, only the environment that it is
deploying to.
This commit fixes the issue by instead using a new env var that is
available that has the node group environment to do a deploy to, so we do
not need to perform any extra lookups