(CDPE-4537) Properly read CD4PE_MODULE_DEPLOY_READ_TIMEOUT env var
Prior to this change, if a user attempted to override the http read
timeout using the `CD4PE_MODULE_DEPLOY_READ_TIMEOUT` env var as an
String, the task would fail, as a string value of the override could not
be applied. When CD4PE 4x was released, it forced the user to define
this override in their kubernetes deployment, forcing the value to be
passed in as a string, meaning the override could never work.

With this change, we modify the code to handle either a string OR an
integer as an override, and fall back to a default of 600 seconds if
no override is detected.