(CDPE-5224) Allow ajax calls to work for all envs
The cd4pe prefix can only be used in environemnts where the UI proxy is
running (i.e. prod), but cannot be used when the proxy is not available
(i.e. backend tests). Since the UI proxy already maps calls to
`/:workspace/ajax` and `/cd4pe/:workspace/ajax` to the backend we can
safely remove the prefix from all ajax calls and have it work for
environments with and without a UI proxy.